pyproject.toml
[project]
name = "workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg==0.30.0",
    "fastapi>=0.116.1",
    "langchain==0.3.26",
    "langchain-chroma>=0.1.2",
    "langchain-community==0.3.26",
    "langchain-google-genai==2.1.8",
    "langchain-openai>=0.3.31",
    "langgraph==0.6.2",
    "pypdf==5.9.0",
    "python-dotenv==1.0.1",
    "redis==6.4.0",
    "squareup==43.0.2.20250820",
    "uvicorn>=0.35.0",
]
